refactor(tui): extract Activity Detail / raw-detail helpers into ui/activity_detail.rs (#4103)
Behavior-preserving move of the Activity-Detail / raw-`v`-detail / pager-text
helper cluster out of the ~13.2k-line ui.rs into a new ui/activity_detail.rs
submodule. No logic changes; visibility widened minimally (pub(super) for
ui.rs/test callers, existing pub(crate) re-exported for mouse_ui/footer_ui).

Implemented with agent assistance.
